Split reflective into a library
