+++ lib/reflective.dart
@@ -2,7 +2,9 @@ library reflective;
 
 import 'dart:mirrors';
 import 'dart:convert';
+import 'package:reflective/src/reflective/reflective.dart';
 
-part 'src/reflective.dart';
+export 'package:reflective/src/reflective/reflective.dart';
+
 part 'src/convert.dart';
 part 'src/optional.dart';
