+++ lib/src/reflective/field_reflection.dart
@@ -1,0 +1,13 @@
+part of reflective.reflective;
+
+abstract class FieldReflection {
+  bool has(Type metadata);
+
+  value(Object entity);
+
+  set(Object entity, value);
+
+  TypeReflection get type;
+
+  String get name;
+}
