+++ lib/src/reflective/type_reflection.dart
@@ -1,8 +1,4 @@
-part of reflective;
-
-type(Type type) => new TypeReflection(type);
-
-instance(Object instance) => new TypeReflection.fromInstance(instance);
+part of reflective.reflective;
 
 class TypeReflection<T> {
   TypeMirror _mirror;
@@ -48,8 +44,7 @@ class TypeReflection<T> {
   bool get isEnum => _mirror is ClassMirror ? (_mirror as ClassMirror).isEnum : false;
 
   List get enumValues {
-    if (!isEnum || _mirror is !ClassMirror)
-      return null;
+    if (!isEnum || _mirror is! ClassMirror) return null;
     return (_mirror as ClassMirror).getField(#values).reflectee;
   }
 
@@ -141,88 +136,4 @@ class TypeReflection<T> {
 
     return library.declarations[new Symbol(name)];
   }
-}
-
-TypeReflection<dynamic> dynamicReflection = new TypeReflection(dynamic);
-
-abstract class FieldReflection {
-  bool has(Type metadata);
-
-  value(Object entity);
-
-  set(Object entity, value);
-
-  TypeReflection get type;
-
-  String get name;
-}
-
-class SimpleFieldReflection extends FieldReflection {
-  Symbol _symbol;
-  VariableMirror _variable;
-  MethodMirror _accessor;
-
-  SimpleFieldReflection(this._symbol, this._variable, this._accessor);
-
-  bool has(Type metadata) =>
-      _variable.metadata.firstWhere((instance) => instance.type.reflectedType == metadata, orElse: () => null) != null;
-
-  value(Object entity) => reflect(entity).getField(_symbol).reflectee;
-
-  set(Object entity, value) => reflect(entity).setField(_symbol, value);
-
-  TypeReflection get type => new TypeReflection.fromMirror(_accessor.returnType);
-
-  String get name => MirrorSystem.getName(_symbol);
-
-  String toString() => name;
-
-  bool operator ==(o) =>
-      o is FieldReflection && _variable.qualifiedName == o._variable.qualifiedName && _symbol == o._symbol;
-}
-
-class TransitiveFieldReflection extends FieldReflection {
-  FieldReflection _source;
-  FieldReflection _target;
-
-  TransitiveFieldReflection(this._source, this._target);
-
-  bool has(Type metadata) => _source.has(metadata);
-
-  value(Object entity) {
-    var sourceValue = _source.value(entity);
-    if (sourceValue == null) return null;
-    return _target.value(sourceValue);
-  }
-
-  set(Object entity, value) => _target.set(_source.value(entity), value);
-
-  TypeReflection get type => _source.type;
-
-  String get name => _source.name + '.' + _target.name;
-}
-
-class Maps {
-  static Map index(Iterable iterable, indexer(key)) {
-    Map result = {};
-    iterable.forEach((i) => result[indexer(i)] = i);
-    return result;
-  }
-
-  static Map where(Map map, predicate(key, value)) {
-    Map result = {};
-    forEach(map, (k, v) {
-      if (predicate(k, v)) result[k] = v;
-    });
-    return result;
-  }
-
-  static forEach(Map map, function(key, value)) {
-    map.keys.forEach((k) => function(k, map[k]));
-  }
-}
-
-class Objects {
-  static int hash(List toHash) =>
-      toHash.fold(17, (e1, e2) => (e1 != null ? e1.hashCode : 1) * 37 + (e2 != null ? e2.hashCode : 1));
 }
